Show description + image with terminal background
Cards with no image will have the header on top and the description as
the card body.

Images, that aren't so wide that they fit in the card will have the
terminal color on the sides. This prevents seeing the previous image
when scrolling.
But there must be a system for redrawing cards without a image.